requirements.txt
# ============================================
# 🛡️ Aegis-G Dependencies
# ============================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0  # Production server

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Graph Database
neo4j==5.14.1

# AI / ML
google-generativeai==0.3.1
numpy==1.24.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Cache & Storage
redis==5.0.1
google-cloud-storage==2.10.0  # Optional for GCS

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
httpx==0.25.2

# Testing (install with: pip install -r requirements.txt[test])
# pytest==7.4.3
# pytest-cov==4.1.0
# pytest-asyncio==0.21.1

# Development (install with: pip install -r requirements.txt[dev])
# ruff==0.1.6
# black==23.11.0

